feat: move applications out to their own file
